requirements.txt
sqlalchemy
psycopg2-binary
APScheduler
requests
anthropic
openai
pdfplumber
python-dotenv
